feat: Replace `hostname` with `origin` for permissions and per dapp selected network state (#17523)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

## **Description**

Currently in Mobile, in-app browser permissions are key'ed incorrectly
by hostname rather than origin. This opens up the small possibility for
hijacking. Origin is the better value to use rather than hostname
because of this. Additionally, Extension has been using origin for
awhile.

This PR fixes that by replacing the usage of hostname with origin in the
BackgroundBridge (which is used by the in-app browser pipeline, WC, and
MM SDK). NOTE that it does not update variable and prop names to match.

There is also a new migration that makes a naive assumption that any
hostnames were made over a secure web connection by prefixing hostname
values with `https://` in the PermissionController and
SelectedNetworkController.

## **Changelog**

<!--
If this PR is not End-User-Facing and should not show up in the
CHANGELOG, you can choose to either:
1. Write `CHANGELOG entry: null`
2. Label with `no-changelog`

If this PR is End-User-Facing, please write a short User-Facing
description in the past tense like:
`CHANGELOG entry: Added a new tab for users to see their NFTs`
`CHANGELOG entry: Fixed a bug that was causing some NFTs to flicker`

(This helps the Release Engineer do their job more quickly and
accurately)
-->

CHANGELOG entry:

## **Related issues**

Fixes:

## **Manual testing steps**

1. Permission a dapp on an https page 
2. Permission a dapp on a http page on a different hostname
3. Build and install this branch
4. The https page should still have permissions as before
5. The http page should not have permissions anymore
6. Everything related to in-app browser, WalletConnect, and MM SDK
connection flow should work the same way as it did before. Use the
buttons on the [test dapp](https://metamask.github.io/test-dapp/) to
trigger these three different flows. Note that the WC and SDK flows
should be triggered from a browser app on the phone NOT the in-app
browser in MetaMask.

## **Screenshots/Recordings**

<!-- If applicable, add screenshots and/or recordings to visualize the
before and after of your change. -->

### **Before**

<!-- [screenshots/recordings] -->

### **After**

<!-- [screenshots/recordings] -->

## **Pre-merge author checklist**

- [ ] I’ve followed [MetaMask Contributor
Docs](https://github.com/MetaMask/contributor-docs) and [MetaMask Mobile
Coding
Standards](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/CODING_GUIDELINES.md).
- [ ] I've completed the PR template to the best of my ability
- [ ] I’ve included tests if applicable
- [ ] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [ ] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.

---------

Co-authored-by: ffmcgee <joao.carlos@consensys.net>
Co-authored-by: Alex Donesky <adonesky@gmail.com>