ci: Bump `setup-ruby` (#19073)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

## **Description**

The version of `setup-ruby` that we were using was not compatible with
macOS 15, causing CI to fail when a runner with a newer OS was chosen.

This PR pins it to the current latest commit on the `v1` branch.