fix: tutorial navigation issues (#19355)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

## **Description**
1. What is the reason for the change?
There were a few issues with tutorial flow. On android there was a
double continue when button was clicked and also we sometimes would
redirect to the last page instead of defaulting to the markets list
view. Also the wallet action button did not have a tutorial condition on
it
2. What is the improvement/solution?
Ensure that we can not double click buttons. Ensure that we are
defaulting to the markets list view on escape routes. Wallet action
button now has a tutorial condition on it when the user is a first time
user
<!--
Write a short description of the changes included in this pull request,
also include relevant motivation and context. Have in mind the following
questions:
1. What is the reason for the change?
2. What is the improvement/solution?
-->

## **Changelog**

<!--
If this PR is not End-User-Facing and should not show up in the
CHANGELOG, you can choose to either:
1. Write `CHANGELOG entry: null`
2. Label with `no-changelog`

If this PR is End-User-Facing, please write a short User-Facing
description in the past tense like:
`CHANGELOG entry: Added a new tab for users to see their NFTs`
`CHANGELOG entry: Fixed a bug that was causing some NFTs to flicker`

(This helps the Release Engineer do their job more quickly and
accurately)
-->

CHANGELOG entry: null

## **Related issues**

Fixes: https://consensyssoftware.atlassian.net/browse/TAT-1378

## **Manual testing steps**

```gherkin
Feature: tutorial navigation fix
  Scenario: user is a first time user and is not geoblocked entering from GTM full screen perps ad
    Given user enters the tutorial from the GTM full screen perps ad

    When user clicks start trading and goes through the flow of tutorial
    Then if they click skip or got it they will go to the markets list

  Scenario: user is a first time user and is geoblocked entering from GTM full screen perps ad
    Given user enters the tutorial from the GTM full screen perps ad

    When user clicks start trading and goes through the flow of tutorial
    Then if they click skip or got it they will go to the markets list

  Scenario: user is a first time user and is not geoblocked and has clicked not now on GTM full screen perps ad
    Given user enters the tutorial from the wallet view perps tab

    When user clicks start trading and goes through the flow of tutorial
    Then if they click skip or got it they will go to the markets list

  Scenario: user is a first time user and is geoblocked and has clicked not now on GTM full screen perps ad
    Given user enters the tutorial from the wallet view perps tab

    When user clicks start trading and goes through the flow of tutorial
    Then if they click skip or got it they will go to the markets list


  Scenario: user is a first time user
    Given user is logged in

    When user clicks on the wallet action button for perps
    Then the tutorial shows
```

## **Screenshots/Recordings**

<!-- If applicable, add screenshots and/or recordings to visualize the
before and after of your change. -->

### **Before**

<!-- [screenshots/recordings] -->

### **After**


https://www.loom.com/share/281f4e93c36843e7abe8ce7ab7482a52?sid=ea06fb96-65ca-412a-b62f-bae023440a3d

<!-- [screenshots/recordings] -->

## **Pre-merge author checklist**

- [x] I’ve followed [MetaMask Contributor
Docs](https://github.com/MetaMask/contributor-docs) and [MetaMask Mobile
Coding
Standards](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/CODING_GUIDELINES.md).
- [x] I've completed the PR template to the best of my ability
- [x] I’ve included tests if applicable
- [x] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [x] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.